Movement Improved (And Camera)

+++ NES-SMB/2DGame/NES-SMB/NES-SMB/Player.h
@@ -21,7 +21,7 @@ public:
 	void setTileMap(TileMap *tileMap);
 	void setPosition(const glm::vec2 &pos);
 	glm::ivec2 getPosition();
-	int getVelocity();
+	float getVelocity();
 	void setVelocity();
 	bool moving();
 	bool moving_up();
